requirements.txt
Babel==2.6.0
Django==2.0.7
django-nested-inline==0.3.7
gunicorn==19.9.0
Pillow==5.2.0
pytz==2018.5
reportlab==3.5.2
sweetify==1.0.3
